backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
elevenlabs
python-multipart
twilio
